DISPATCH-1017 Add console to make build
